+++ server/lib/tuist/command_events/postgres.ex
@@ -50,7 +50,7 @@ defmodule Tuist.CommandEvents.Postgres do
   def get_command_events_by_name_git_ref_and_project(%{name: name, git_ref: git_ref, project: %Project{id: project_id}}) do
     Repo.all(
       from(e in Event,
-        where: e.name == ^name and e.git_ref == ^git_ref and e.project_id == ^project_id,
+        where: e.name == ^name and like(e.git_ref, ^git_ref) and e.project_id == ^project_id,
         select: e
       )
     )